Include paths

+++ src/common/nullpo.c
@@ -6,6 +6,8 @@
 
 #include "nullpo.h"
 
+#include "common/showmsg.h"
+
 #include <stdio.h>
 #include <stdarg.h>
 #include <stdlib.h>
@@ -13,8 +15,6 @@
 #ifdef __GNUC__
 #include <execinfo.h>
 #endif
-
-#include "../common/showmsg.h"
 
 struct nullpo_interface nullpo_s;
 
@@ -43,11 +43,11 @@ void assert_report(const char *file, int line, const char *func, const char *tar
 	ShowError("--- %s --------------------------------------------\n", title);
 	ShowError("%s:%d: '%s' in function `%s'\n", file, line, targetname, func);
 #ifdef __GNUC__
-	size = backtrace (array, 10);
-	strings = backtrace_symbols (array, size);
+	size = (int)backtrace(array, 10);
+	strings = backtrace_symbols(array, size);
 	for (i = 0; i < size; i++)
 		ShowError("%s\n", strings[i]);
-	free (strings);
+	free(strings);
 #endif
 	ShowError("--- end %s ----------------------------------------\n", title);
 }
